Refactoring SDK to allow authentication using a private token as well as OAuth2 workflow.
This commit is not backwards compatible and requires PHP 7.1+ from now
on due to type hinting and return values. While the MedialabService
class is backwards compatible, the config object has been rewritten to
allow 2 different kinds of config, and each config loads their own
client to execute the HTTP requests.
This is so that the OAuth2 workflow still uses the League\OAuth2 client
library, while the PrivateToken workflow uses the GuzzleHTTP directly.